feat: get created parts on loft
