+++ src/onpy/features/extrude.py
@@ -44,36 +44,7 @@ class Extrude(Feature):
 
     def get_created_parts(self) -> list[Part]:
         """Gets a list of the parts this feature created"""
-
-        script = dedent(f"""
-            function(context is Context, queries) {{
-                var query = qCreatedBy(makeId("{self.id}"), EntityType.BODY);
-
-                return transientQueriesToStrings( evaluateQuery(context, query) ); 
-            }}
-            """)
-        
-        response = self._client._api.endpoints.eval_featurescript(
-            document_id=self._partstudio.document.id,
-            version=WorkspaceWVM(self._partstudio.document.default_workspace.id),
-            element_id=self._partstudio.id,
-            script=script,
-            return_type=model.FeaturescriptResponse,
-        )
-        
-        part_ids_raw = unwrap(
-            response.result, message="Featurescript failed get parts created by feature"
-        )["value"]
-
-        part_ids = [i["value"] for i in part_ids_raw]
-
-        available_parts = self._api.endpoints.list_parts(
-            document_id=self._partstudio.document.id,
-            version=WorkspaceWVM(self._partstudio.document.default_workspace.id),
-            element_id=self._partstudio.id
-        )
-
-        return [Part(self._partstudio, part) for part in available_parts if part.partId in part_ids]
+        return self._get_created_parts_inner()
 
     @property
     @override
